requirements.txt
PyQt6>=6.9.0
requests>=2.32.0
Pillow>=11.3.0
python-magic>=0.4.27
openai>=1.0.0
dashscope>=1.14.0
